+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.12] - 2026-04-23
+
+### Docs
+- Update README.md
+
+### Other
+- Update api/policy/rules.yaml
+- Update godot/.gitignore
+- Update godot/Dockerfile
+- Update godot/Makefile
+- Update godot/caddy.conf
+- Update godot/docker-compose.yml
+- Update godot/scripts/build.sh
+- Update godot/scripts/install.sh
+- Update godot/scripts/run-bundle.sh
+- Update godot/scripts/validate-all.sh
+- ... and 3 more files
+
 ## [0.1.11] - 2026-04-23
 
 ### Docs